backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LangChain & LLM
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Vector Store
qdrant-client>=1.12.0
langchain-qdrant>=0.2.0

# Knowledge Graph
neo4j>=5.25.0
langchain-neo4j>=0.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0

# Document Storage
minio>=7.2.0

# Document Parsing
pypdf>=5.0.0
unstructured[all-docs]>=0.16.0
python-multipart>=0.0.12

# Utilities
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=24.4.0
